+++ src/main/java/com/treinamento/model/PeriodoCurso.java
@@ -9,7 +9,7 @@ public class PeriodoCurso {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Double id;
 
     @Column(name = "periodo_curso", nullable = false)
     private String periodo;
@@ -29,11 +29,11 @@ public class PeriodoCurso {
         this.periodo = periodo;
     }
 
-    public Long getId() {
+    public Double getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Double id) {
         this.id = id;
     }
